+++ apps/python/register/utils/mleh.py
@@ -5,10 +5,17 @@ import sys
 from typing import Union
 from lm_eval import  utils
 from lm_eval.tasks import TaskManager, include_path, initialize_tasks
+from lm_eval.utils import eval_logger, positional_deprecated, simple_parse_args_string
+from lm_eval.tasks import TaskManager, get_task_dict
+from lm_eval.evaluator_utils import run_task_tests
 
-from utils.uploader import get_ConfigurableTask 
+from typing import List, Optional, Union
+
+
 from utils.sql import create_dataset_table, register_task, create_task_table, checked_task
 import psycopg2
+
+
 
 
 def parse_eval_args() -> argparse.Namespace:
@@ -185,3 +192,101 @@ def cli_register_task(args: Union[argparse.Namespace, None] = None) -> None:
             eval_logger.info(f"Task {task_name_i} registered successfully")
         else:
             eval_logger.info(f"Task {task_name_i} already registered")
+
+
+# cutted def simple_evaluate(..) from lm-eval-harness to generate config task commit:7d9922c80114218eaf43975b7655bb48cda84f50
+@positional_deprecated
+def get_ConfigurableTask(
+    tasks: Optional[List[Union[str, dict, object]]] = None,
+    num_fewshot: Optional[int] = None,
+    check_integrity: bool = False,
+    gen_kwargs: Optional[str] = None,
+    task_manager: Optional[TaskManager] = None,
+    verbosity: str = "INFO",
+    predict_only: bool = False,
+
+):
+    """Instantiate and evaluate a model on a list of tasks.
+
+    :param tasks: list[Union[str, dict, Task]]
+        List of task names or Task objects. Task objects will be taken to have name task.EVAL_HARNESS_NAME if defined and type(task).__name__ otherwise.
+    :param num_fewshot: int
+        Number of examples in few-shot context
+    :param check_integrity: bool
+        Whether to run the relevant part of the test suite for the tasks
+    :param gen_kwargs: str
+        String arguments for model generation
+        Ignored for all tasks with loglikelihood output_type
+    :param predict_only: bool
+        If true only model outputs will be generated and returned. Metrics will not be evaluated
+
+    :return
+        Task dictionary
+    """
+    eval_logger.setLevel(getattr(logging, f"{verbosity}"))
+
+    seed_message = []
+
+    if seed_message:
+        eval_logger.info(" | ".join(seed_message))
+
+    if tasks is None:
+        tasks = []
+    if len(tasks) == 0:
+        raise ValueError(
+            "No tasks specified, or no tasks found. Please verify the task names."
+        )
+
+    if gen_kwargs is not None:
+        gen_kwargs = simple_parse_args_string(gen_kwargs)
+        eval_logger.warning(
+            "generation_kwargs specified through cli, these settings will update set parameters in yaml tasks. "
+            "Ensure 'do_sample=True' for non-greedy decoding!"
+        )
+        if gen_kwargs == "":
+            gen_kwargs = None
+
+    if task_manager is None:
+        task_manager = TaskManager(verbosity)
+
+    eval_logger.info(
+        "get_task_dict has been updated to accept an optional argument, `task_manager`"
+        "Read more here:https://github.com/EleutherAI/lm-evaluation-harness/blob/main/docs/interface.md#external-library-usage"
+    )
+    task_dict = get_task_dict(tasks, task_manager)
+    for task_name in task_dict.keys():
+        task_obj = task_dict[task_name]
+        if isinstance(task_obj, tuple):
+            _, task_obj = task_obj
+            if task_obj is None:
+                continue
+
+        if task_obj.get_config("output_type") == "generate_until":
+            if gen_kwargs is not None:
+                task_obj.set_config(
+                    key="generation_kwargs", value=gen_kwargs, update=True
+                )
+
+        if predict_only:
+            log_samples = True
+            eval_logger.info(
+                f"Processing {task_name} in output-only mode. Metrics will not be calculated!"
+            )
+            # we have to change the class properties post-hoc. This is pretty hacky.
+            task_obj.override_metric(metric_name="bypass")
+
+        if num_fewshot is not None:
+            if (default_num_fewshot := task_obj.get_config("num_fewshot")) == 0:
+                eval_logger.info(
+                    f"num_fewshot has been set to 0 for {task_name} in its config. Manual configuration will be ignored."
+                )
+            else:
+                eval_logger.warning(
+                    f"Overwriting default num_fewshot of {task_name} from {default_num_fewshot} to {num_fewshot}"
+                )
+                task_obj.set_config(key="num_fewshot", value=num_fewshot)
+
+    if check_integrity:
+        run_task_tests(task_list=tasks)
+
+    return task_dict
